pyproject.toml
[tool.poetry]
name = "theme"
version = "0.1.0"
description = "Theme switcher for nvim and alacritty"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.scripts]
theme = "theme.theme:main"

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.3.2"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
nose = "^1.3.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
